Update Strings.xml
New lines added.  Since I work I could not see if there were new lines
-----------------------
In the translators I have placed myself first for the moment. It is up to you to delete the old translators or leave them (since I do not see movement of them.)
--------------------
By last. Sorry for my English. It's a bit of mine and of the translator.