Move request time out "forever" to end of array
Move the 'forever' option to the end of the request time out list.  This serves the purpose of preventing users from accidentally authorizing applications or the `su` binary indefinitely which may have unintended long-term effects.  By moving the 'forever' option to the end it also orders the list by length of time and makes 'once' the first option which allows for fast acceptance.